requirements.txt
Django>=5.2,<6.1
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
django-crispy-forms>=2.0,<3.0
crispy-bootstrap5>=2024.2,<2026.0
python-dotenv>=1.0,<2.0
whitenoise>=6.5,<7.0
gunicorn>=22.0,<24.0
pypdf>=4.0,<6.0
reportlab>=4.0,<5.0
Pillow>=10.4,<12.0
django-allauth[mfa,socialaccount]>=65.0,<66.0
keel @ git+https://github.com/okeefedaniel/keel.git@3b978b1949dfae729e38dd39b2627ddf9aa2c872
